moves testing functions into a separate convex/testing forlder
